requirements.txt
pandas
scikit-learn
rank-bm25
sentence-transformers
flask
flask-cors
requests
python-dotenv
